cambios al registro y mejora de beta de login

+++ registro.php
@@ -1,17 +1,19 @@
 <?php
-if($_POST){
-  $db = file_get_contents("usuarios.json");
-  $usuarios = json_decode($db, true);
- $usuarios[] = [
-     "email" =>$_POST["email"],
-     "usuario" =>$_POST["usuario"],
-     "password" =>password_hash ($_POST["password"], PASSWORD_DEFAULT),
-     "passwordConfirm" =>password_hash ($_POST["passwordConfirm"], PASSWORD_DEFAULT)
- ];
-  $db = json_encode($usuarios);
-  file_put_contents('usuarios.json',$db);
+if ($_POST) {
+  if ($_POST["password"] == $_POST["passwordConfirm"]) {
+    $db = file_get_contents("usuarios.json");
+    $usuarios = json_decode($db, true);
+    $usuarios[] = [
+      "email" => $_POST["email"],
+      "usuario" => $_POST["usuario"],
+      "password" => password_hash($_POST["password"], PASSWORD_DEFAULT),
+    ];
+    $db = json_encode($usuarios);
+    file_put_contents('usuarios.json', $db);
+  } else { }
 }
 ?>
+
 
 <!DOCTYPE html>
 <html lang="en">
@@ -28,9 +30,7 @@ if($_POST){
   <link rel="stylesheet" href="style.css">
   <script src="https://unpkg.com/ionicons@4.5.10-0/dist/ionicons.js"></script>
 </head>
-
 <body>
-
   <header>
     <nav class="fixed-top navbar navbar-expand-lg navbar-light bg-light">
       <a class="navbar-brand" style="font-family: 'Alata', sans-serif;" href="index.html">
@@ -71,11 +71,11 @@ if($_POST){
       </div>
       <div class="form-group">
         <label for="password">Password</label>
-        <input name="password" type="password" class="form-control" id="password" placeholder="Password" required>
+        <input name="password" onchange="if(this.checkValidity()) form.passwordConfirm.pattern = this.value;" type="password" class="form-control" id="password" placeholder="Password" required>
       </div>
       <div class="form-group">
         <label for="passwordConfirm">Repite la contraseña</label>
-        <input name="passwordConfirm" type="password" class="form-control" id="passwordConfirm" placeholder="Password" required>
+        <input name="passwordConfirm" onchange="this.setCustomValidity(this.validity.patternMismatch ? 'Las contraseñas no coinciden' : '');" type="password" class="form-control" id="passwordConfirm" placeholder="Password" required>
       </div>
       <button type="submit" class="btn btn-outline-dark">Crear cuenta</button>
     </form>
@@ -132,12 +132,14 @@ if($_POST){
     </div>
   </footer>
   <!-- Scripts -->
+
   <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous">
   </script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js" integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1" crossorigin="anonymous">
   </script>
   <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous">
   </script>
+
 </body>
 
 </html>
